[Movie] removed duplicated timestamp code
Event already has "startDate". Not need for "timestamp".

+++ app/src/main/java/org/xvdr/robotv/service/DataService.java
@@ -268,7 +268,7 @@ public class DataService extends Service {
             name += movie.getTitle();
         }
 
-        return timer.create(movie.getChannelUid(), movie.getStartTime(), (int)movie.getDuration(), name);
+        return timer.create(movie.getChannelUid(), movie.getStartTime(), movie.getDuration(), name);
     }
 
     public Connection getConnection() {
